# perf: Multiple `StyledText` optimizations